feat: implement new footer component and navigation structure
