--- components/navbar/NavbarClientWrapper.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import NavbarIntl from "./NavbarIntl";
-
-export default function NavbarClientWrapper(props: { className?: string }) {
-  return <NavbarIntl className={props.className} />;
-}
