+++ data/navigations.ts
@@ -1,0 +1,12 @@
+export const navLinks = {
+  pages: {
+    homepage: "/",
+    features: "/features",
+    profile: "/profile",
+  },
+  company: {
+    aboutus: "/about",
+    term: "/terms",
+    blog: "/blog",
+  },
+} as const;
