+++ features/home/components/HomeComp.tsx
@@ -1,13 +1,13 @@
-import FooterComp from "@/components/footer/FooterComp";
 import Container from "./banner/Container";
 import ContentComp from "./content/ContentComp";
+import Footer from "@/components/container/footer";
 
 const HomeComp = () => {
   return (
     <div>
       <Container />
       <ContentComp />
-      <FooterComp />
+      <Footer />
     </div>
   );
 };
